feat: support floating-point values for tab top margin settings
- Change `tab_top_margin` type to `f32` in `UiConfig` and `AppState` to allow precision adjustments.
- Use `0.25` steps when incrementing/decrementing with `[-]` and `[+]` buttons or keys.
- Update text input handling to accept digit and decimal point keys.
- Encode float value as bits in `SaveTabTopMargin` enum action to support standard Eq/PartialEq derives.
- Update layout computation in `src/ui.rs` to round the floating-point margin to the nearest cell row for TUI rendering.
- Update settings and layout unit tests to verify floating-point adjustments and saving.

+++ src/app/config_io.rs
@@ -129,7 +129,7 @@ impl App {
         }
     }
 
-    pub(super) fn save_tab_top_margin(&mut self, val: i16) {
+    pub(super) fn save_tab_top_margin(&mut self, val: f32) {
         if self.update_config_file("tab top margin", |content| {
             crate::config::upsert_section_value(content, "ui", "tab_top_margin", &val.to_string())
         }) {
